# build/resources/main/static/swagger.yaml
openapi: 3.0.3
info:
  title: Swagger Horizonal BANK - OpenAPI 3.0
  description: |-
    This is a Hotrizonal financial application Server based on the OpenAPI 3.0 specification.  A Horizonal Bank financial application is being developed to streamline and simplify various banking functionalities for users. The application will allow users to transfer funds, view transaction history, manage digital cards associated with their accounts, and automatically create an account upon registration and login. This application aims to provide a seamless and efficient experience for users, leveraging modern technologies and design practices to achieve these goals.
    
    Some useful links:
    - [The horizonal bank app repository](https://github.com/isaac-zik-rgb/HorizonalBank/tree/master)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: http://swagger.io/terms/
  contact:
    email: okechukwuisaac111@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: https://horizonal-mz29tyzk.b4a.run/api
  - url: https://horizonalbank.onrender.com/api
  - url: http://localhost:8080/api
...
paths:
  /auth/register:
    post:
      tags:
        - auth
      summary: register a new user
      description: create a new user based on the provided infomation
      operationId: addUser
      requestBody:
        description: create a new user with its details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                  description: first name of the new user
                last_name:
                  type: string
                  description: last name of the new user
                password:
                  type: string
                  description: Password of the new user, must be 8 character long, contains a capital letter, small letter, digit and a symbol
                email:
                  type: string
                  description: email of the new user. must be a valid email
                address:
                  type: string
                  description: The address of the user
                date_of_birth:
                  type: string
                  description: The date of birth of the user
                postal_code:
                  type: number
                  description: Postal code of the new user
                state:
                  type: string
                  description: The state of the new user
                ssn:
                  type: string
                  description: SSN of the new user
      
      responses:
        200:
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        400:
          description: Bad request, check your values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user:
    get:
      tags:
        - user
      summary: Return a user by the token
      description: return the login user data
      operationId: userId
      security:
        - BearerAuth: []
      responses:
        200:
          description: user successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/login:
    post:
      tags:
        - auth
      summary: login a user
      description: this route handles user login request
      operationId: loginUser
      requestBody:
        description: login details that is need to be given
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: email of the already existed user
                password:
                  type: string
                  description: password of the already existed user
            
      responses:
        200:
          description: User successfully loged in
          content:
           application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    description: the user Id
                  token:
                    type: string
                    description: Jwt token for the user, which will be used for subsequent request
      
        404:
          description: No User found with the provided email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        
      
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/validate:
    get:
      tags:
        - auth
      summary: validate user token
      description: this route validate a token and make sure its vaild
      operationId: validateToken
      security:
        - BearerAuth: []
      parameters:
        - name: token
          in: query
          required: true
          description: The token to be validated
          schema:
            type: string
      
      responses:
        200:
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    description: Whether the token is valid or not
                  userId:
                    type: string
                    description: The ID of the user associated with the token

        400:
          description: Bad request, invalid token format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/email/verify:
    get:
      tags:
        - auth
      summary: verifies that the given email belongs to a user in the database, this route is only used for password reset
      description: Finds a user who's email equals with the given email
      operationId: verifyEmail
      parameters:
        - name: email
          in: query
          required: true
          description: A string email that will use to verify if the user is in the database
          schema:
            type: string
      responses:
        200:
          description: An ok that returns a string saying the email is valid
          content:
            application/json:
              schema:
                type: string
  /auth/password/reset:
    put:
      tags:
        - auth
      summary: A put request that will take in an email, new-password and confirm-password in order to rest the user password
      description: Reset a user password
      operationId: resetId
      requestBody:
        description: The body that will contain the email, new-password and confirm-password
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: The email of an existing user
                new_password:
                  type: string
                  description: The new password of the user
                confirm_password:
                  type: string
                  description: confirm password of the user, will be use to validate the new password to get a match
      responses:
        200:
          description: An ok that shows that the password have been modified
          content:
            application/json:
              schema:
                type: string
        404:
          description: No User found by the provided email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: A detailed message explaining the error
        code:
          type: integer
          description: HTTP status code of the error
        details:
          type: object
          description: Additional details about the error
    User:
      type: object
      properties:
                first_name:
                  type: string
                  description: first name of the new user
                last_name:
                  type: string
                  description: last name of the new user
                password:
                  type: string
                  description: Password of the new user, must be 8 character long, contains a capital letter, small letter, digit and a symbol
                email:
                  type: string
                  description: email of the new user. must be a valid email
                address:
                  type: string
                  description: The address of the user
                date_of_birth:
                  type: string
                  description: The date of birth of the user
                postal_code:
                  type: number
                  description: Postal code of the new user
                state:
                  type: string
                  description: The state of the new user
                ssn:
                  type: string
                  description: SSN of the new user